+++ src/attr/_funcs.py
@@ -164,7 +164,13 @@ def assoc(inst, **changes):
         be found on *cls*.
     :raise attr.exceptions.NotAnAttrsClassError: If *cls* is not an ``attrs``
         class.
+
+    ..  deprecated:: 17.1.0
+        Use :func:`evolve` instead.
     """
+    import warnings
+    warnings.warn("assoc is deprecated and will be removed after 2018/01.",
+                  DeprecationWarning)
     new = copy.copy(inst)
     attrs = fields(inst.__class__)
     for k, v in iteritems(changes):
@@ -176,3 +182,40 @@ def assoc(inst, **changes):
             )
         _obj_setattr(new, k, v)
     return new
+
+
+def evolve(inst, **changes):
+    """
+    Create a new instance, based on *inst* with *changes* applied.
+
+    :param inst: Instance of a class with ``attrs`` attributes.
+    :param changes: Keyword changes in the new copy.
+
+    :return: A copy of inst with *changes* incorporated.
+
+    :raise attr.exceptions.AttrsAttributeNotFoundError: If *attr_name* couldn't
+        be found on *cls*.
+    :raise attr.exceptions.NotAnAttrsClassError: If *cls* is not an ``attrs``
+        class.
+
+    ..  versionadded:: 17.1.0
+    """
+    cls = inst.__class__
+    for a in fields(cls):
+        attr_name = a.name  # To deal with private attributes.
+        if attr_name[0] == "_":
+            init_name = attr_name[1:]
+            if attr_name not in changes:
+                changes[init_name] = getattr(inst, attr_name)
+            else:
+                # attr_name is in changes, it needs to be translated.
+                changes[init_name] = changes.pop(attr_name)
+        else:
+            if attr_name not in changes:
+                changes[attr_name] = getattr(inst, attr_name)
+    try:
+        return cls(**changes)
+    except TypeError as exc:
+        k = exc.args[0].split("'")[1]
+        raise AttrsAttributeNotFoundError(
+            "{k} is not an attrs attribute on {cl}.".format(k=k, cl=cls))